requirements.txt
# 滕王阁序量化交易系统依赖
akshare>=1.10.0
tushare>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0

# Supabase 数据库
supabase>=2.0.0
python-dotenv>=1.0.0

